Finished big refactoring/simplification of the way file input and output parameters are represented in Osiris::Job (and associated XML files). Treating 'upload' params in their own set of methods was baked into the Job object, which was bad - it put Dancer dependencies into the data objects, and made it annoyingly buggy to test the Job from test scripts.
Now it's the job of the calling program (whether it's a Dancer web app, or
a test script, or something else down the track) to copy input files into
the job's working directory.  Once that's done, the input files are just
parameters (with the value being the fully qualified path to the working
directory's copy of the file.)

This makes everything a lot less messy.

Passes unit tests, and the web app and ptah work with it too.