Issue #6 - error handling
Added basic error handling to ptah.pl and the web app.  The sig_child event
in ptah now checks the child process's return value, and if it's not 0,
sets the job's status to 'error' rather than 'done'.
Fixed Osiris::Job so that the 'error' status sets the finished timestamp.
Added styles for failed jobs in the joblist, and fixed the job.tt template
so that failed jobs display the print.prt file - this is where the
actual error message will be.