Issue #1 - Cancel button Issue #8 - Smarter Back button behaviour
Added some javascript to app.tt to make the Cancel button work, and which
also checks whether the user has touched any of the form inputs and pops
a confirm dialogue if they have.

The Cancel button and the Back links on job pages now go back to the
document.referrer, which takes care of the #8 smarter back button
behaviour.