Issue #2 - Customisable extra fields
Added the extra-metadata-fields feature.  Configuration is an xml file
in the same format as the Isis api xml.  The form is generated by the
same code and template (Osiris::Form and form.tt).  There's a bit
of javascript to make the form fold up and down.   The extra metadata
fields are stored in the job xml file where they can be picked up
by CoataGlue.